+++ client/apps/game/src/ui/components/settlement/use-canvas-interactions.ts
@@ -1,12 +1,19 @@
-import { useCallback, useRef } from "react";
-import { SettlementLocation } from "./settlement-types";
+import { useCallback, useEffect, useRef, useState } from "react";
+import {
+  MAX_ZOOM_LEVEL,
+  MAX_ZOOM_RANGE,
+  MIN_ZOOM_LEVEL,
+  MIN_ZOOM_RANGE,
+  MINIMAP_HEIGHT,
+  MINIMAP_WIDTH,
+  SETTLEMENT_CENTER,
+} from "./settlement-constants";
+import { MapViewState, SettlementLocation } from "./settlement-types";
+import { normalizedToContractCoords } from "./settlement-utils";
 
 interface CanvasInteractionsProps {
   availableLocations: SettlementLocation[];
   settledLocations: SettlementLocation[];
-  mapCenter: { x: number; y: number };
-  mapSize: { width: number; height: number };
-  setMapCenter: (center: { x: number; y: number }) => void;
   setSelectedLocation: (location: SettlementLocation | null) => void;
   onSelectLocation: (location: SettlementLocation) => void;
 }
@@ -17,9 +24,6 @@ interface CanvasInteractionsProps {
 export const useCanvasInteractions = ({
   availableLocations,
   settledLocations,
-  mapCenter,
-  mapSize,
-  setMapCenter,
   setSelectedLocation,
   onSelectLocation,
 }: CanvasInteractionsProps) => {
@@ -28,6 +32,118 @@ export const useCanvasInteractions = ({
   const isDraggingRef = useRef(false);
   const lastMousePositionRef = useRef<{ x: number; y: number } | null>(null);
   const mouseStartPositionRef = useRef<{ x: number; y: number } | null>(null);
+
+  // Map view state
+  const [mapViewState, setMapViewState] = useState<MapViewState>({
+    mapCenter: { x: SETTLEMENT_CENTER, y: SETTLEMENT_CENTER },
+    mapSize: { width: MINIMAP_WIDTH, height: MINIMAP_HEIGHT },
+    zoomLevel: 1,
+  });
+
+  // Interaction state
+  const [customNormalizedCoords, setCustomNormalizedCoords] = useState({ x: 0, y: 0 });
+
+  const mapCenter = mapViewState.mapCenter;
+  const mapSize = mapViewState.mapSize;
+  const zoomLevel = mapViewState.zoomLevel;
+
+  const setMapCenter = (center: { x: number; y: number }) =>
+    setMapViewState((prev) => ({ ...prev, mapCenter: center }));
+
+  // Function to set zoom level and update map size
+  const setZoom = useCallback((zoomOut: boolean, delta = 10) => {
+    setMapViewState((prev) => {
+      // Current range
+      const currentRange = prev.mapSize.width;
+
+      // Check zoom limits
+      if (!zoomOut && currentRange < MIN_ZOOM_RANGE) return prev;
+      if (zoomOut && currentRange > MAX_ZOOM_RANGE) return prev;
+
+      // Calculate new size
+      const ratio = MINIMAP_WIDTH / MINIMAP_HEIGHT;
+      const deltaX = Math.round(delta * ratio);
+      const deltaY = delta;
+
+      // Update zoom level for UI feedback
+      const newZoomLevel = zoomOut
+        ? Math.max(MIN_ZOOM_LEVEL, prev.zoomLevel - 0.05)
+        : Math.min(MAX_ZOOM_LEVEL, prev.zoomLevel + 0.05);
+
+      return {
+        ...prev,
+        zoomLevel: newZoomLevel,
+        mapSize: {
+          width: prev.mapSize.width + (zoomOut ? 2 * deltaX : -2 * deltaX),
+          height: prev.mapSize.height + (zoomOut ? 2 * deltaY : -2 * deltaY),
+        },
+      };
+    });
+  }, []);
+
+  // Function to zoom to a specific level
+  const zoomToLevel = useCallback((targetZoomLevel: number) => {
+    setMapViewState((prev) => {
+      const ratio = MINIMAP_WIDTH / MINIMAP_HEIGHT;
+
+      // Calculate target width based on zoom level (0.5 to 2)
+      // 0.5 = zoomed out (MAX_ZOOM_RANGE)
+      // 2 = zoomed in (MIN_ZOOM_RANGE)
+      const targetWidth = MAX_ZOOM_RANGE - ((targetZoomLevel - 0.5) / 1.5) * (MAX_ZOOM_RANGE - MIN_ZOOM_RANGE);
+
+      // Ensure we stay within bounds
+      const newWidth = Math.max(MIN_ZOOM_RANGE, Math.min(MAX_ZOOM_RANGE, targetWidth));
+      const newHeight = newWidth / ratio;
+
+      return {
+        ...prev,
+        zoomLevel: targetZoomLevel,
+        mapSize: {
+          width: newWidth,
+          height: newHeight,
+        },
+      };
+    });
+  }, []);
+
+  // Apply zoom level when it changes
+  useEffect(() => {
+    zoomToLevel(mapViewState.zoomLevel);
+  }, [mapViewState.zoomLevel, zoomToLevel]);
+
+  // Center map on custom coordinates and zoom in
+  const centerOnCoordinates = useCallback(() => {
+    const contractCoords = normalizedToContractCoords(customNormalizedCoords.x, customNormalizedCoords.y);
+
+    // Set map center to the input coordinates
+    setMapViewState((prev) => ({
+      ...prev,
+      mapCenter: contractCoords,
+      zoomLevel: 2,
+    }));
+  }, [customNormalizedCoords]);
+
+  // Reset map to center
+  const resetMapCenter = useCallback(() => {
+    setMapViewState({
+      mapCenter: { x: SETTLEMENT_CENTER, y: SETTLEMENT_CENTER },
+      mapSize: { width: MINIMAP_WIDTH, height: MINIMAP_HEIGHT },
+      zoomLevel: 1,
+    });
+    setCustomNormalizedCoords({ x: 0, y: 0 });
+  }, []);
+
+  // Handle coordinate input changes
+  const handleCoordinateChange = useCallback((e: React.ChangeEvent<HTMLInputElement>, coord: "x" | "y") => {
+    // Allow empty string, negative sign, or valid numbers
+    if (e.target.value === "" || e.target.value === "-" || /^-?\d+$/.test(e.target.value)) {
+      const value = e.target.value === "" || e.target.value === "-" ? e.target.value : parseInt(e.target.value);
+      setCustomNormalizedCoords((prev) => ({
+        ...prev,
+        [coord]: value,
+      }));
+    }
+  }, []);
 
   // Function to get current hovered location (for external access if needed)
   const getHoveredLocation = useCallback(() => {
@@ -257,5 +373,14 @@ export const useCanvasInteractions = ({
     handleMouseDown,
     handleMouseUp,
     getHoveredLocation,
+    mapCenter,
+    mapSize,
+    setMapCenter,
+    setZoom,
+    resetMapCenter,
+    handleCoordinateChange,
+    customNormalizedCoords,
+    centerOnCoordinates,
+    zoomLevel,
   };
 };
